Fix a long unrepoted bug that prevent shuttles from transforming
Somehow, when I grabbed from SOS2 repo and recompiled that myself, the compiler optimized the code in CompBecomePawn.myPawn() so that the local variable No.0 is typeof Pawn, but not typeof PawnKindDef. This update fixes the problem that others using github version as well as keep the compatibility between my own compiled version and github version. After all, the IL code that's loading the pawn variable must be 3 instructions earlier than that stores the apparel